[Temp] Add extra duel GTs to compensate for MAX_STRING_CHARS limit
Currently, the MAX_STRING_CHARS limit only allows a string to use up to 1024 characters. As a result, map lists/pools larger than that are not supported. By using an additional gametype, we can use an additional gametype config. This allows us to split the map list/pool across the versions to reduce the string length.